#1891: extract interfaces schema subtree into schema_interfaces.go
Second increment of the #1891 domain split. Pure code motion: the
"interfaces" top-level subtree of setSchema (including its #1319 PR-3
typed-leaf doc comment) moves verbatim into schema_interfaces.go as
schemaInterfaces, together with the tunnelSchemaChildren() and
wireguardSchemaNode() constructors it composes (both are referenced
only from the interfaces subtree).

No desc/validator/structure changes. Full canonical node-inventory
dump (1,962 nodes, all schemaNode fields) diffed before/after:
byte-identical. go build ./pkg/config clean.

Part of #1891.
